Functionalize BaseGoalScreen.
