--- src/goals/DefaultGoal/BaseGoalScreen/index.tsx
@@ -1,26 +1,0 @@
-import { StoreState } from "../../../types";
-import { connect } from "react-redux";
-import { RouteComponentProps } from "react-router";
-import { LocalizeContextProps } from "react-localize-redux";
-import { Goal, GoalProps } from "../../../types/goals";
-import BaseGoalScreen from "./BaseGoalScreen";
-
-export interface TParams {
-  id: string;
-}
-
-export function mapStateToProps(
-  state: StoreState,
-  ownProps: GoalProps & RouteComponentProps<TParams> & LocalizeContextProps
-): GoalProps {
-  let idNumber: number = parseInt(ownProps.match.params.id);
-  let goal: Goal | undefined;
-  if (!Number.isNaN(idNumber)) {
-    goal = state.goalsState.historyState.history[idNumber];
-  }
-  return {
-    goal: goal,
-  };
-}
-
-export default connect(mapStateToProps)(BaseGoalScreen);
