+++ src/goals/DefaultGoal/DisplayProg/displayProg.tsx
