+++ src/goals/DefaultGoal/tests/BaseGoalWidget.test.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import { User } from "../../../../types/user";
-import { Goal } from "../../../../types/goals";
-import { CreateCharInv } from "../../../CreateCharInv/CreateCharInv";
+import { Goal } from "../../../types/goals";
+import { User } from "../../../types/user";
+import { CreateCharInv } from "../../CreateCharInv/CreateCharInv";
 import BaseGoalWidget from "../BaseGoalWidget";
 
 it("renders without crashing", () => {
   const div = document.createElement("div");
-  const user: User = new User("TestUser", "TestUsername", "TestPass");
+  const user = new User("TestUser", "TestUsername", "TestPass");
   const goal: Goal = new CreateCharInv();
   goal.user = user;
   ReactDOM.render(<BaseGoalWidget goal={goal} />, div);
